gemini_screen_agent/requirements.txt
torch
transformers
pillow
pyttsx3
keyboard
